lf review: lfd: implement daemon and move agent commands
Replace maestro web UI with Unix socket daemon. Agent commands moved from `lf agent` to `lfd` binary. Daemon manages agent lifecycle via socket protocol instead of SQLite polling.
